+++ src/main/java/us/kbase/auth2/service/api/APIToken.java
@@ -1,7 +1,6 @@
 package us.kbase.auth2.service.api;
 
 import us.kbase.auth2.lib.token.StoredToken;
-import us.kbase.auth2.lib.identity.MfaStatus;
 import us.kbase.auth2.service.common.ExternalToken;
 
 public class APIToken extends ExternalToken {
@@ -9,25 +8,14 @@ public class APIToken extends ExternalToken {
 	//TODO JAVADOC or swagger
 	
 	private final long cachefor;
-	private final MfaStatus mfa;
 	
 	public APIToken(final StoredToken token, final long tokenCacheTimeMillis) {
 		super(token);
 		cachefor = tokenCacheTimeMillis;
-		mfa = token.getMfa();
 	}
 
 	public long getCachefor() {
 		return cachefor;
-	}
-	
-	/**
-	 * Gets the MFA authentication status for this token.
-	 * 
-	 * @return the MFA authentication status.
-	 */
-	public MfaStatus getMfa() {
-		return mfa;
 	}
 
 	@Override
@@ -35,7 +23,6 @@ public class APIToken extends ExternalToken {
 		final int prime = 31;
 		int result = super.hashCode();
 		result = prime * result + (int) (cachefor ^ (cachefor >>> 32));
-		result = prime * result + ((mfa == null) ? 0 : mfa.name().hashCode());
 		return result;
 	}
 
@@ -49,13 +36,6 @@ public class APIToken extends ExternalToken {
 		}
 		APIToken other = (APIToken) obj;
 		if (cachefor != other.cachefor) {
-			return false;
-		}
-		if (mfa == null) {
-			if (other.mfa != null) {
-				return false;
-			}
-		} else if (!mfa.equals(other.mfa)) {
 			return false;
 		}
 		return true;
